# docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.RefreshRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
...
host: localhost:8080
info:
  contact:
    email: your-email@example.com
    name: API Support
    url: https://github.com/your-repo/gin-fast/issues
  description: 基于Gin框架的快速开发脚手架API文档
  license:
    name: MIT
    url: https://github.com/your-repo/gin-fast/blob/master/LICENSE
  termsOfService: https://github.com/your-repo/gin-fast
  title: Gin-Fast API
  version: "1.0"
paths:
...
  /captcha/image:
    get:
      description: 根据验证码ID获取验证码图片
      parameters:
      - description: 验证码ID
        in: query
        name: captchaId
        required: true
        type: string
      - default: 130
        description: 图片宽度
        in: query
        name: width
        type: integer
      - default: 30
        description: 图片高度
        in: query
        name: height
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回验证码图片
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取验证码图片
      tags:
      - 认证
  /config/get:
    get:
      consumes:
      - application/json
      description: 获取系统配置信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回配置信息
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取配置信息
      tags:
      - 配置管理
  /config/update:
    put:
      consumes:
      - application/json
      description: 更新系统配置信息
      parameters:
      - description: 配置信息
        in: body
        name: config
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功更新配置信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新配置信息
      tags:
      - 配置管理
  /login:
    post:
      consumes:
      - application/json
      description: 用户登录获取访问令牌
      parameters:
      - description: 登录请求参数
        in: body
        name: loginReq
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回访问令牌
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 用户名或密码错误
          schema:
            additionalProperties: true
            type: object
      summary: 用户登录
      tags:
      - 认证
  /refreshToken:
    post:
      consumes:
      - application/json
      description: 使用刷新令牌获取新的访问令牌
      parameters:
      - description: 刷新令牌请求参数
        in: body
        name: refreshReq
        required: true
        schema:
          $ref: '#/definitions/models.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回新的访问令牌
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 刷新令牌无效或过期
          schema:
            additionalProperties: true
            type: object
      summary: 刷新访问令牌
      tags:
      - 认证
...
  /users/logout:
    post:
      consumes:
      - application/json
      description: 用户登出，撤销access token和refresh token
      produces:
      - application/json
      responses:
        "200":
          description: 成功登出
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 用户未登录
          schema:
            additionalProperties: true
            type: object
      summary: 用户登出
      tags:
      - 认证
...
swagger: "2.0"